backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6

# Redis for pub/sub and cache
redis==5.2.1

# Authentication
PyJWT==2.10.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1

# HTTP client (for Ollama and external APIs)
httpx==0.28.1

# Password hashing (for production)
passlib[bcrypt]==1.7.4

# Environment variables
python-dotenv==1.0.1

# Rate limiting
slowapi==0.1.9

# Development tools
pytest==8.3.4
pytest-asyncio==0.25.2
